async version of notificationcenter delegate
Revert "Completion handler notification delegate instead of async version (#945)"
This reverts commit 1e2375de89477363ef02b02daa23bd2247d40883.
